// packages/typespec-test/test/Fleet.Management/generated/typespec-ts/src/models/models.ts
export interface ArmResource extends ArmResourceBase {
  /** Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName} */
  readonly id: string;
  /** The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts" */
  readonly type: string;
  /** Azure Resource Manager metadata containing createdBy and modifiedBy information. */
  readonly systemData?: SystemData;
}
// ...
export interface ProxyResourceBase extends ArmResource {}
// ...
export type ProvisioningState = string;
// ...
export interface Deployment extends ProxyResourceBase {
  /** The resource-specific properties for this resource. */
  properties?: DeploymentProperties;
}

/** The properties of deployment */
export interface DeploymentProperties {
  /** Deployment ID */
  deploymentId?: string;
  /** Images deployed */
  deployedImages?: Image[];
  /** Deployment date UTC */
  readonly deploymentDateUtc?: Date;
  /** The status of the last operation. */
  readonly provisioningState?: ProvisioningState;
}

/** An image resource belonging to a catalog resource. */
export interface Image extends ProxyResourceBase {
  /** The resource-specific properties for this resource. */
  properties?: ImageProperties;
}
// ...
export interface DeploymentListResult {
  /** The Deployment items on this page */
  value: Deployment[];
  /** The link to the next page of items */
  nextLink?: string;
}